backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
django-redis>=5.4.0
gunicorn>=21.2.0,<22.0
